feat(bundle): Consolidate exported bundle format
These changes ensure that any `Gymnast` export can also be imported as:

```js
import Grid from 'gymnast/src/grid'
```

The format is `gymnast/src/${export}` where export is the name of the export
but with the first letter always in lower case
